+++ extensions/git/src/gitEditor.ts
@@ -2,6 +2,7 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
+import * as path from 'path';
 import { TabInputText, Uri, window, workspace } from 'vscode';
 import { IIPCHandler, IIPCServer } from './ipc/ipcServer';
 import { EmptyDisposable, IDisposable } from './util';
@@ -14,7 +15,7 @@ export class GitEditor implements IIPCHandler {
 
 	private disposable: IDisposable = EmptyDisposable;
 
-	constructor(ipc?: IIPCServer) {
+	constructor(private ipc?: IIPCServer) {
 		if (ipc) {
 			this.disposable = ipc.registerHandler('git-editor', this);
 		}
@@ -35,6 +36,21 @@ export class GitEditor implements IIPCHandler {
 				});
 			});
 		}
+	}
+
+	getEnv(): { [key: string]: string } {
+		const env: { [key: string]: string } = {
+			VSCODE_GIT_EDITOR_NODE: process.execPath,
+			VSCODE_GIT_EDITOR_EXTRA_ARGS: (process.versions['electron'] && process.versions['microsoft-build']) ? '--ms-enable-electron-run-as-node' : '',
+			VSCODE_GIT_EDITOR_MAIN: path.join(__dirname, 'git-editor-main.js')
+		};
+
+		const config = workspace.getConfiguration('git');
+		if (config.get<boolean>('useEditorAsCommitInput')) {
+			env.GIT_EDITOR = `"${path.join(__dirname, this.ipc ? 'git-editor.sh' : 'git-editor-empty.sh')}"`;
+		}
+
+		return env;
 	}
 
 	dispose(): void {
